[BACK-338] Update token `Session` struct
Use the new token `Session` struct as modified in `shoreline` as
part of BACK-338.
Although the migration has already been run, and will not be run again,
this code has been modified so that `Session` is changed to reflect
the current reality in `shoreline`.
This keeps it consistent for any developer in future relying on `Session`
to be correct.

* Fixes BACK-338