[BACK-1572] Migrate or archive duplicate `jellyfish` Omnipod data
Although [BACK-1379] fixed the `deviceId` field with additional
apostrophes in Omnipod data, it did not migrate the `id` and `_id`
fields, which are required for correct deduplication and/or archival.

Since the `deviceId` field has already removed the apostrophes,
this new migration script finds duplicates by generating the expected
hash and comparing it against Mongo, and either migrates or archives
duplicates accordingly.

* Fixes [BACK-1572]