Add app validation routes and service.
This commit adds app validation routes that will generate a challenge for the
user to use in verification. App Validation is the process of determining
whether an app is a valid instance of a mobile app. This involves two parts.
Attestation to confirm an app is a valid instance and Assertion for the app to
assert legitimacy in exchange for sensitive or private information.