[BACK-2164] Add attestation verification route. (#618)
* Add attestation verification route under POST /v1/attestations/verifications.

This commit continues the attestation API to include verification of an
attestation. Once a user generates an attestation challenge, they must verify
the attestation object received from Apple which includes this challenge. Once
verified, the user will be able to make assertions in exchange for a secret.

* Add app assertion verification route.

This commit adds app assertion verification. The user can get an assertion
verified only after being attested and only after generating an assertion
challenge. Once the assertion is verified, the user will be granted access to
some private or sensitive information.