add a check to care team alerting endpoints for user permissions
Users shouldn't be modifying any alert configs if they don't have the proper
permissions to receive alerts in the first place.

It's possible that allowing them to delete, even if they don't have alerting
permission could be useful, but we'll deal with that if/when some client wants
that ability.

Part of BACK-2501