alerts endpoints read followedID from request path
With the recently accomplished separation of the alerts.Alert, the alerts HTTP
handler functions shift to read the followedID not from the message body, but
rather from the request path.

With the removal of the possibility of a userID existing in a message body, a
number of tests cases are no longer necessary, as those potential errors have
been eliminated. They can no longer occur. \o/

BACK-2681