add endpoints for storing push notification device tokens
The mongodb collection maintains a unique index on the combination of the
userID and some key value for the token itself (in this case a SHA-256
checksum hash). This allows for a simple upsert whenever a key is received,
without having to check for duplicates or worry about appending to a list or
anything like that. If the received key is a duplicate, it will replace the old
one. Otherwise a new one is inserted.

While only Apple device tokens are supported now, care has been taken to allow
for other manufacturers' tokens in the future. In theory, that would mean
adding a property to the DeviceToken struct, presumably "Android". Then
modifying the DeviceToken's key() method to produce a key from either the
Apple token, or Android token, whichever is found. Validation should ensure
that only one or the other exists.

BACK-2506