[BACK-2681] improvements to alert config endpoints (#664)
* improved alert config JSON marshaling

- uses the proper case for "enabled"
- omits repeat if empty

BACK-2681

* add a JSON marshaler for DurationMinutes

In testing hydrophone interactions it became apparent that DurationMinutes
would need to be properly marshaled as well as unmarshaled.

Without this, a DurationMinutes value of 5 minutes would be marshaled as a
bare time.Duration, i.e. 300000000000.

BACK-2681

* accept a service client for alerts endpoints

Since calls from other services might come through using service credentials,
they should be accepted. This changes the parameters of authorization checks,
as one can't simply use the userID passed with the token, as there is none.

BACK-2681

* ensure consistency between userID values in alerts endpoints

With three possible sources for a userID in alerts endpoints, there's no clear
way to determine which value should have priority in the event of a
disagreement. Therefore, for safety, if there is any disagreement in the
values, return an error.

BACK-2681

* extract the core of alerts into an inner struct

Now code that reads alerts configs (i.e. HTTP handler methods) can read an
Alerts struct, thereby preventing a user supplied userID or followedID
slipping in by accident and overwriting someone else's data.

BACK-2681

* alerts endpoints read followedID from request path

With the recently accomplished separation of the alerts.Alert, the alerts HTTP
handler functions shift to read the followedID not from the message body, but
rather from the request path.

With the removal of the possibility of a userID existing in a message body, a
number of tests cases are no longer necessary, as those potential errors have
been eliminated. They can no longer occur. \o/

BACK-2681

* adds validation of repeat values in alerts configs

BACK-2500

* adds validation of delay and threshold values in alerts configs

Validation values are taken from the docs (see Jira) as of this day.

BACK-2500

* rename normUnits => normalizeUnitsCase for clarity

BACK-2500

* Updates to care partner alerts stemming from code review

https://github.com/tidepool-org/platform/pull/664

+ followedID => followedUserID
+ Don't validate exact threshold values, just use the min and max from glucose
+ other minor fixes

BACK-2500

* alerts endpoints: extract path knowledge from helper

checkUserIDConsistency is further simplified by removing knowledge of the path
parameters.

BACK-2500