adapt the data service's request handlers to rest.HandlerFunc
When I started working on the alerts configs, I had a discussion with Todd
about routing, and how the data service for some unknown reason uses its own
authentication method, rather than the various api.Require* methods provided
by platform and used in the other services.

In the process of working on https://github.com/tidepool-org/platform/pull/671
it became clear to me that the reason was the difference in the function
signatures of request handlers in the data service as opposed to those in
other services. Specifically, data service request handlers have this
signature:

    func Foo(service.Context)

Whereas rest.HandlerFunc has this signature:

    func Foo(rest.ResponseWriter, *rest.Request)

So I modified the data service's MakeRoute function to accept a list of
rest.MiddlewareSimple, and added a ToRestRoute method that when provided with
the existing adapter function in the data api service, would convert the data
service routes to *rest.Route, while also applying middleware.

The result is that data service routes can be specified including
rest.MiddlewareSimple (like api.Require), and we no longer require a separate
Authentication helper for data service routes.

There are some TODO comments that indicate that going further down the data
service handler rabbit hole might not have been the intended path, but rather
moving to a rest.HandlerFunc signature was at one time the plan (in 2018 or
so). If so, this should only make it clearer what needs to change to effect
that migration.