add ConfigLoader interface
This interface provides a way to migrate away from the env.reporter. The
intention is to use envconfig in future work.

I tried implementing the existing config.Reporter interface with envconfig,
but wasn't able to find a solution that felt good. I wrote a solution using
reflection, but it felt overcomplicated (as anything using reflection does),
and I tried a brute force approach, but it was heavily coupled to the config
structs, and was brittle as a result.

This solution uses an inversion of responsibility pattern. Unfortunately it
gets a bit repetitive with all the config/client/platform terminology that
already exists, but each individual piece at least is simple and
straightforward. Hopefully in time, we can further simplify or even just
better name some of the pieces to let the code read more clearly.

BACK-2500