[BACK-2737] adds an alerts client (#679)
* remove Timeout from config.Config

Removing the timeout simplifies the Config and related Clients which no longer
need to be concerned with it. It's used by a single client request. That
request now sets a timeout on the context, preventing a bunch of unnecessary
complication everywhere else.

* deprecated functions are deprecated

* add ConfigLoader interface

This interface provides a way to migrate away from the env.reporter. The
intention is to use envconfig in future work.

I tried implementing the existing config.Reporter interface with envconfig,
but wasn't able to find a solution that felt good. I wrote a solution using
reflection, but it felt overcomplicated (as anything using reflection does),
and I tried a brute force approach, but it was heavily coupled to the config
structs, and was brittle as a result.

This solution uses an inversion of responsibility pattern. Unfortunately it
gets a bit repetitive with all the config/client/platform terminology that
already exists, but each individual piece at least is simple and
straightforward. Hopefully in time, we can further simplify or even just
better name some of the pieces to let the code read more clearly.

BACK-2500

* adds an alerts client

BACK-2500

* make client config's UserAgent property optional

The value is useful for debugging, but it's not being set shouldn't stop a
test or any other process from running.

Also fixed a few calls to Go stdlib deprecated functions.

* update mongo/travis config

Stolen from Alex's good work in https://github.com/tidepool-org/platform/pull/681

* force snyk to use govendor as its package manager

Its autodetection isn't working correctly.