set CreatedUserID to the id of the user making the request
This is intended to help other tools like ORCA be able to find basic
information about the person responsible for uploading the data set.

The test isn't the prettiest, but given that there were 0 tests in this
package, and there was a lot of work required to set up the test scaffolding,
I figured this was a better starting place than having no tests. It can be
improved over time.

BACK-2814