response inspector errors don't abort their request
Previously a response inspector that returned an error would fail the entire
request call, even if the HTTP response was 200. On top of that, it would
prevent later response inspectors from running, which could cause other
problems, for example inaccurate metrics.

If a response inspector wishes to take action based on an error it encounters,
it should handle that itself.

BACK-2820