break AddBin down into smaller functions
Namely, addBinAfter, addBinBefore, and replaceBin.

I found this breakdown a little easier to reason about, as it separates the
if-then structures and exposes what I feel is a more obvious, and more
task-oriented flow.

It might also be interesting too, to have a "readjust" function, that given a
slice of buckets, simply re-orders them and inserts/removes gaps as
necessary. Then AddBin just becomes something like:

    buckets = append(buckets, newBucket)
    readjustBuckets(buckets)

The signature for removeExcessBuckets has an extra argument, because I
couldn't quite figure out how to make the compiler happy without it.