adds ability to send push notifications to Apple devices
The intent is to allow sending push notifications, while providing minimal
friction to switching push notification providers. A future requirement to
send notifications to Android is expected, but in the meantime, the front-end
team has settled on APNs for their notifications. It's simple enough to
support multiple notification providers, and this interface can easily be
removed in the future if we decide on a single provider and don't feel that
the abstraction brings further value.

The change brings in a new dependency, github.com/sideshow/apns2, however it's
a small library, with very few dependencies of its own[1], primarily
those packages I would likely use in implementing the functionality myself.

[1]: Aside from the kingpin library, whose functionality we aren't using
anyway, as it's only used for a standalone binary that's part of the package,
it's not used by the library itself.

BACK-2507