don't look for device data provenance without a JWT (part 2)
In a previous commit (see BACK-2995), empty tokens were excluded, but there
can be tokens that aren't JWTs, which would still trigger a spurious log
message. This change removes more cases where a token exists, but isn't a
JWT. The goal is to log a warning when we expect a JWT, but don't get one.

BACK-3027