add a "go-test" Makefile recipe and extract common variables
This recipe can be called to run the ginkgo test suite via `go test`, which
provides caching. Be warned however, there are a few tests that appear to be
flappy when so doing. Further investigation is needed.

This common variable extraction allows them to be modified from the CLI as
well, e.g. to pass a label filter.