rename the task/test package to task/tasktest
This has a few follow-on effects. Namely one no longer needs to have a special
name for the task/test package (previously this was custom named to
"taskTest"). Now it just follows naturally as "tasktest" with no custom naming
required.

This change brings the name inline with current Go best practices for package
names and package naming, and also makes it easier for LSP-enabled editors to
automatically manage Go imports. In addition, it follows patterns found in the
Go stdlib (see httptest, slogtest, fstest, iotest, etc.)

I also renamed task/test/mock.go to task/tasktest/taskclient_mock_gen.go just
to make it extra clear that the file is generated, as well as what it
contains (not just a mock, but a mock of the task client).

[^1]: https://go.dev/doc/effective_go#package-names