fix for a test that fails under `go test`
There has been discussion of running our test suites via `go test` instead of
`ginkgo` to get the package-level parallelism that `go test` offers (and
caching!), without having to re-work all of our tests to make use of ginkgo
v2's parallelism. To this end, the Makefile previously had a `go-test` recipe
added. However, that recipe routinely fails one test in particular.

The test in question was found to be timing-out when run via `go test`. I
suspect it's because under `go test` it's run in parallel, and isn't getting
as much running time as a result. In any event, giving it more time allows the
test to complete successfully (ironic, given that using `go test` is overall
faster, at least for me).

I'm not sure that using `go test` is a long-term win over using ginkgo's
parallelism though, as we can only access the package-level parallelism of `go
test`, that is, we can't call `testing.T.Parallel()` to get test-level
parallelism from `go test`, because Ginkgo doesn't support it, so we're back
to re-working our tests to use Ginkgo's parallelism.

But... at least now we can run the tests both ways and compare.