make the shared store more obvious; improve cleanup
Due to a misunderstanding I had about ginkgo test suites, I had added the
DeferredCleanup mechanism, but I now see that isn't required and so have
removed it.

By putting the shared store behind a function, we make it explicit when it's
being used, so there's a hint of when and how one should clean up after
themselves.

This opens more opportunities to easily reuse this shared store in other
tests, which will come in follow-on commits.