minimize database timeouts for tests
These were the two tests that took the longest to run.

Each test was waiting five seconds for a timeout that we fully intended to
happen. By shortening the timeout, we don't need to wait as long.

I'm not sure why we're really testing this, because I don't see how MongoDB
will somehow in the future connect to hosts that don't exist.

This chops 10 seconds off from each run.