minimal modifications of the deprecated service tests, to test the new service
The AuthClient test removal is the only thing I find in any way controversial
here. That test requests the auth client, then calls a method on it, to ensure
that it returns what is expected. However, since the new service doesn't
initialize that auth client itself, rather, it's just returning whatever is
passed to it via a setter, there's no point in testing this here.