run both ci-test and go-ci-test via Travis.
This is intended as a trial to ensure that both make ci-test and make
go-ci-test find the same errors. The latter is significantly faster, so it
would be preferable to use it instead, but of course we need to ensure that
we're not running things incorrectly or missing out on something provided by
the current ginkgo-based solution.

For more info/discussion:
  - https://tidepoolteam.slack.com/archives/CF59ADWLB/p1727967169327199
  - https://github.com/tidepool-org/platform/pull/777