tweak how private is handled
This uses the PRIVATE variable a little closer to what
matters. Previously, one had to follow the PRIVATE variable to the
setting of REPOSITORY_SUFFIX, then follow that its use in setting
REPOSITORY_NAME, then trace that all around to understand the effect
of PRIVATE.

This way, we read the value of PRIVATE, and use it all in the same
place, and we don't need REPOSITORY_SUFFIX at all.